Move media queries into an array scale Also re use media helper in syntax highlighting theme part
