+++ src/theme/parts/breakpoints.ts
@@ -1,7 +1,3 @@
 import { Breakpoints } from '../../types/theme/Breakpoints';
 
-export const breakpoints: Breakpoints = {
-  small: '640px',
-  medium: '768px',
-  large: '1024px',
-};
+export const breakpoints: Breakpoints = ['640px', '768px', '1024px'];
